Start refactoring locid_transform
