+++ components/locid_transform/README.md
@@ -20,7 +20,7 @@ This minimize method returns a new Locale that is the result of running the
 ## Examples
 
 ```rust
-use icu_locid_transform::{CanonicalizationResult, LocaleCanonicalizer};
+use icu_locid_transform::{TransformResult, LocaleCanonicalizer};
 use icu_locid::Locale;
 
 let provider = icu_testdata::get_provider();
@@ -31,40 +31,40 @@ let mut locale: Locale = "ja-Latn-fonipa-hepburn-heploc"
     .expect("parse failed");
 assert_eq!(
     lc.canonicalize(&mut locale),
-    CanonicalizationResult::Modified
+    TransformResult::Modified
 );
 assert_eq!(locale.to_string(), "ja-Latn-alalc97-fonipa");
 ```
 
 ```rust
-use icu_locid_transform::{CanonicalizationResult, LocaleCanonicalizer};
+use icu_locid_transform::{TransformResult, LocaleCanonicalizer};
 use icu_locid::Locale;
 
 let provider = icu_testdata::get_provider();
 let lc = LocaleCanonicalizer::new(&provider).expect("create failed");
 
 let mut locale: Locale = "zh-CN".parse().expect("parse failed");
-assert_eq!(lc.maximize(&mut locale), CanonicalizationResult::Modified);
+assert_eq!(lc.maximize(&mut locale), TransformResult::Modified);
 assert_eq!(locale.to_string(), "zh-Hans-CN");
 
 let mut locale: Locale = "zh-Hant-TW".parse().expect("parse failed");
-assert_eq!(lc.maximize(&mut locale), CanonicalizationResult::Unmodified);
+assert_eq!(lc.maximize(&mut locale), TransformResult::Unmodified);
 assert_eq!(locale.to_string(), "zh-Hant-TW");
 ```
 
 ```rust
-use icu_locid_transform::{CanonicalizationResult, LocaleCanonicalizer};
+use icu_locid_transform::{TransformResult, LocaleCanonicalizer};
 use icu_locid::Locale;
 
 let provider = icu_testdata::get_provider();
 let lc = LocaleCanonicalizer::new(&provider).expect("create failed");
 
 let mut locale: Locale = "zh-Hans-CN".parse().expect("parse failed");
-assert_eq!(lc.minimize(&mut locale), CanonicalizationResult::Modified);
+assert_eq!(lc.minimize(&mut locale), TransformResult::Modified);
 assert_eq!(locale.to_string(), "zh");
 
 let mut locale: Locale = "zh".parse().expect("parse failed");
-assert_eq!(lc.minimize(&mut locale), CanonicalizationResult::Unmodified);
+assert_eq!(lc.minimize(&mut locale), TransformResult::Unmodified);
 assert_eq!(locale.to_string(), "zh");
 ```
