+++ components/locid_transform/benches/locale_canonicalizer.rs
@@ -7,10 +7,11 @@ mod helpers;
 use criterion::{criterion_group, criterion_main, Criterion};
 use icu_locid::Locale;
 use icu_locid_transform::LocaleCanonicalizer;
+use icu_locid_transform::LocaleExpander;
 
 fn canonicalize_bench(c: &mut Criterion) {
     let provider = icu_testdata::get_provider();
-    let lc = LocaleCanonicalizer::new(&provider).unwrap();
+    let lc = LocaleCanonicalizer::try_new_unstable(&provider).unwrap();
 
     let mut group = c.benchmark_group("canonicalize");
 
@@ -37,7 +38,7 @@ fn canonicalize_bench(c: &mut Criterion) {
 
 fn canonicalize_noop_bench(c: &mut Criterion) {
     let provider = icu_testdata::get_provider();
-    let lc = LocaleCanonicalizer::new(&provider).unwrap();
+    let lc = LocaleCanonicalizer::try_new_unstable(&provider).unwrap();
 
     let mut group = c.benchmark_group("canonicalize-noop");
 
@@ -66,7 +67,7 @@ fn canonicalize_noop_bench(c: &mut Criterion) {
 
 fn maximize_bench(c: &mut Criterion) {
     let provider = icu_testdata::get_provider();
-    let lc = LocaleCanonicalizer::new(&provider).unwrap();
+    let lc = LocaleExpander::try_new_unstable(&provider).unwrap();
 
     let mut group = c.benchmark_group("likelysubtags");
